Create extensible WeeklyActivityBase class
