+++ App/Containers/Styles/WeeklyActivityBaseStyle.js
